Move inventory/item stuff to character_inventory.cpp
Including:
- Character::handle_contents_changed
- Character::count_softwares
- Character::max_single_item_length
- Character::max_single_item_volume
- Character::best_pocket
- Character::try_add
- Character::i_add
- Character::i_at
- Character::i_rem
- Character::i_rem_keep_contents
- Character::i_add_or_drop
- Character::recur_internal_locations
- Character::all_items_loc
- Character::top_items_loc
- Character::invlet_to_item
- Character::get_item_position
- Character::drop
- Character::allocated_invlets
- Character::has_active_item
- Character::can_drop
- Character::drop_invalid_inventory
